Last patch via home at winter.
Refactored battle and enemy cards, new enemyCard view, new enemies sprites and cards: Rats. Test add of Zenject, first attempt to integrate this frameword to project.

+++ Assets/Scripts/Cards/SituationCards/Event/PropertyEvents/PoisonEvent.cs
@@ -7,9 +7,9 @@ namespace Cards.SituationCards.Event.PropertyEvents
     [CreateAssetMenu(menuName = "Event/PoisonEvent", fileName = "PoisonEvent")]
     public class PoisonEvent : PropertyEventDef
     {
-        [SerializeField] private float _poisonModifier;
+        [SerializeField] private int _poisonModifier;
 
-        public float PoisonModifier => _poisonModifier;
+        public int PoisonModifier => _poisonModifier;
         
         public override void Accept(IPropertyVisitor visitor)
         {
